Allow docs/Makefile to be portable
Currently docs/Makefile downloads Conda but only the Linux version.
This commit allows it to detect the operating system and download the
corresponding Conda installer.

Additionally, `make` uses the sh shell which does not support the `source`
command. This causes the commands that use `source` to fail on some
systems that do not have the sh shell linked to something else (say dash).

Signed-off-by: Daniel Lim Wee Soong <weesoong.lim@gmail.com>